Some compile fixes and refactoring.
git-svn-id: svn://svn.berlios.de/gale/trunk@348 2d798ef8-6e06-0410-941a-9092aa770104

+++ gale/include/gale/math/color.h
@@ -120,7 +120,7 @@ struct ColorModel
 
         if (S>0) {
             H*=6;
-            long long i=roundToZero(H);
+            long i=roundToZero(static_cast<float>(H));
             double f=H-i;
 
             double p=V*(1-S);
